pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "perpetuals-trading-dep"
version = "0.2.0"
description = "Perpetuals trading tooling and analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=4.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["src"]
